wells_fargo_agent/pyproject.toml
[project]
name = "wells-fargo-agent"
version = "0.1.0"
description = "WFAP Wells Fargo Bank Agent - Credit evaluation with JWT signing"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk>=0.3.0",
    "google-adk>=1.8.0",
    "google-genai>=1.27.0",
    "pydantic>=2.11.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.0",
    "PyJWT>=2.8.0",
    "cryptography>=41.0.0",
    "litellm>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
